Many changes to bring closer to support of spring-social 1.1.4
Merge changes from Alfresco/spring-social-salesforce
-Unable to fully use Salesforce Sandbox
--Remove hardcoded userinfo url in useroperations
--Add the ability to pass in the gateway url for the authorization
server used in userinfo
-getBlob is returning a closed InputStream
--override ByteArrayHttpMessageConverter to allow all media types
--change extractor to use the byteArray message converter in the
RestRemplate
--to keep the existing method signature as InputStream return byteArray
as byteArrayInputStream
--remove default content-type header used in other tests to allow return
on non json content.
-Update Salesforce API version used
--Add method to set version to be used by method calls
--Add validation to version provided
--Default version is now v37.0
--Refactor where version and default version are located (ApiTemplate
and ApiOperations
--Update all uses of API_VERSION to now use getVersion method
--Update and Add Tests for version string
--Add getServices method that does not require version to be passed but
uses the current set version
--When the version number passed to setVersion is invalid (format,
blank, null) we no longer silently fail but throw
InvalidSalesforceApiVersionException.
-JsonParseException when response is BAD_OAUTH_TOKEN
--Bad_OAuth_Token is returned as a single string not json as expected
from other errors
--Add code to look at jsonParseException to test if it is actually this
error. Return this error in the proper format to be handled as a known
error. This will allow so that when (if) Salesforce fixes this issue a
new release will not be needed.
--Response handling suffers drom the same end of stream close issue
after single use.  Buffer the stream as a byteArray so that it can be
used in error handling and additional testing after an exception is
thrown.
--When Bad_OAuth_Token is found it is returned as
InvalidAuthorizationException instead of as an
InsufficientPermissionException.  The former should be recoverable with
a token refresh (uses a refresh token).  The latter should indicate a
need to reauthenticate and possible approve additional
permissions/scopes for the user.

Update/remove dependencies to those required for 1.1.4
Refactor any errors or changes based on dependency changes
Update author and license header copyright year